remove method "d" from global namespace
ruby_parser_extras defines a method "d" (which is used exactly once, in the same file) in the public namespace (main). This conflicts with Wrong's "d" method which is part of a module; for some reason even when I "include Wrong::D" the ruby_parser one is called. It's basically bad form for a gem to clutter the global namespace in any case.

This patch puts it inside the (only) object where it's called.